Move Person to v3/domain
